Added api monitoring

+++ openai/src/main/java/com/mardillu/openai/model/requests/LogApiRequest.kt
@@ -2,8 +2,8 @@ package com.mardillu.openai.model.requests
 
 data class LogApiRequest(
     val route: String,
-    val request_time: Int,
-    val response_time: Int,
-    val duration: Int,
+    val request_time: Long,
+    val response_time: Long,
+    val duration: Long,
     val response_code: Int,
 )
